Actualizacion Imagenes

+++ controladores/ordenes.controlador.php
@@ -513,7 +513,7 @@ MOSTRAR ORDENES PARA SUMAR DEL ASESOR
 
 
 
-				$rutaPortada = "../vistas/img/default/default.png";
+				$rutaPortada = "";
 
 
 
@@ -647,7 +647,7 @@ MOSTRAR ORDENES PARA SUMAR DEL ASESOR
 
 
 
-				$rutaFotoPrincipal = "../vistas/img/default/default.png";
+				$rutaFotoPrincipal = "";
 
 
 
@@ -845,9 +845,9 @@ MOSTRAR ORDENES PARA SUMAR DEL ASESOR
 
 					"totalOrden" => $datos["totalOrden"],
 
-					"imgPortada" => substr($rutaPortada, 3),
+					"imgPortada" => $rutaPortada !== "" ? substr($rutaPortada, 3) : "",
 
-					"imgFotoPrincipal" => substr($rutaFotoPrincipal, 3),
+					"imgFotoPrincipal" => $rutaFotoPrincipal !== "" ? substr($rutaFotoPrincipal, 3) : "",
 
 					"fecha_ingreso" => $fechaDeIngreso,
 
@@ -1061,7 +1061,9 @@ MOSTRAR ORDENES PARA SUMAR DEL ASESOR
 
 
 
-					unlink("../" . $datos["antiguaFotoPortada"]);
+					if (!empty($datos["antiguaFotoPortada"]) && $datos["antiguaFotoPortada"] !== "vistas/img/default/default.png") {
+						unlink("../" . $datos["antiguaFotoPortada"]);
+					}
 
 
 
@@ -1207,7 +1209,9 @@ MOSTRAR ORDENES PARA SUMAR DEL ASESOR
 
 
 
-					unlink("../" . $datos["antiguaFotoPrincipal"]);
+					if (!empty($datos["antiguaFotoPrincipal"]) && $datos["antiguaFotoPrincipal"] !== "vistas/img/default/default.png") {
+						unlink("../" . $datos["antiguaFotoPrincipal"]);
+					}
 
 
 
@@ -1403,9 +1407,9 @@ MOSTRAR ORDENES PARA SUMAR DEL ASESOR
 
 					"totalOrdenEditar" => $datos["totalOrdenEditar"],
 
-					"imgPortada" => substr($rutaPortada, 3),
+					"imgPortada" => $rutaPortada !== "" ? substr($rutaPortada, 3) : "",
 
-					"imgFotoPrincipal" => substr($rutaFotoPrincipal, 3),
+					"imgFotoPrincipal" => $rutaFotoPrincipal !== "" ? substr($rutaFotoPrincipal, 3) : "",
 
 					"seleccionarPedido" => $datos["seleccionarPedido"],
